+++ BACK/to-do/src/main/java/com/example/todo/Todo/Services/TodoService.java
@@ -40,7 +40,7 @@ public class TodoService {
 
     }**/
 
-    public void removeTask(int id){
+    public void removeTask(Long id){
         if(!taskRepository.existsById(id)){
             throw new IllegalStateException("Task with id: "+id+" doesn't exist and can't be remove");
         }
@@ -52,7 +52,7 @@ public class TodoService {
     } **/
 
     @Transactional //Spring will manage the transaction (the update in this case)
-    public void updateTask(int id,String title, String description, LocalDate date, String taskState, String taskPriority){
+    public void updateTask(Long id,String title, String description, LocalDate date, String taskState, String taskPriority){
         Task task = taskRepository.findById(id).orElseThrow(() -> new IllegalStateException("Task with id: "+id+" doesn't exist and can't be remove"));
 
         if(title != null
